+++ packages/spacecat-shared-ahrefs-client/test/index.test.js
@@ -47,16 +47,37 @@ describe('AhrefsAPIClient', () => {
     ],
   };
 
-  const topPagesResponse = [
-    {
-      url: 'page-url-1',
-      sum_traffic: 100,
-    },
-    {
-      url: 'page-url-2',
-      sum_traffic: 300,
-    },
-  ];
+  const topPagesResponse = {
+    pages: [
+      {
+        url: 'page-url-1',
+        sum_traffic: 100,
+      },
+      {
+        url: 'page-url-2',
+        sum_traffic: 300,
+      },
+    ],
+  };
+
+  const organicTrafficMetricsResponse = {
+    metrics: [
+      {
+        date: '2024-01-29T00:00:00Z',
+        org_traffic: 179364,
+        paid_traffic: 72,
+        org_cost: 11284251,
+        paid_cost: 2675,
+      },
+      {
+        date: '2024-02-05T00:00:00Z',
+        org_traffic: 176236,
+        paid_traffic: 52,
+        org_cost: 10797893,
+        paid_cost: 1724,
+      },
+    ],
+  };
 
   before('setup', function () {
     this.clock = sandbox.useFakeTimers({
@@ -183,7 +204,7 @@ describe('AhrefsAPIClient', () => {
           'sum_traffic',
         ].join(','),
         where: JSON.stringify(filter),
-        order_by: 'sum_traffic_merged',
+        order_by: 'sum_traffic',
         date,
         target,
         limit: specifiedLimit,
@@ -199,7 +220,7 @@ describe('AhrefsAPIClient', () => {
       const result = await client.getTopPages(target, specifiedLimit);
       expect(result).to.deep.equal({
         result: topPagesResponse,
-        fullAuditRef: `https://example.com/site-explorer/top-pages?select=url%2Csum_traffic&order_by=sum_traffic_merged&date=${date}&target=${target}&limit=${specifiedLimit}&mode=prefix&output=json&where=%7B%22and%22%3A%5B%7B%22field%22%3A%22sum_traffic%22%2C%22is%22%3A%5B%22gt%22%2C0%5D%7D%5D%7D`,
+        fullAuditRef: `https://example.com/site-explorer/top-pages?select=url%2Csum_traffic&order_by=sum_traffic&date=${date}&target=${target}&limit=${specifiedLimit}&mode=prefix&output=json&where=%7B%22and%22%3A%5B%7B%22field%22%3A%22sum_traffic%22%2C%22is%22%3A%5B%22gt%22%2C0%5D%7D%5D%7D`,
       });
     });
   });
@@ -237,6 +258,32 @@ describe('AhrefsAPIClient', () => {
       expect(result).to.deep.equal({
         result: backlinksResponse,
         fullAuditRef: `https://example.com/site-explorer/all-backlinks?select=title%2Curl_from%2Curl_to&order_by=domain_rating_source%3Adesc%2Ctraffic_domain%3Adesc&target=test-site.com&limit=${upperLimit}&mode=prefix&output=json&where=%7B%22and%22%3A%5B%7B%22field%22%3A%22is_dofollow%22%2C%22is%22%3A%5B%22eq%22%2C1%5D%7D%2C%7B%22field%22%3A%22is_content%22%2C%22is%22%3A%5B%22eq%22%2C1%5D%7D%2C%7B%22field%22%3A%22domain_rating_source%22%2C%22is%22%3A%5B%22gte%22%2C29.5%5D%7D%2C%7B%22field%22%3A%22traffic_domain%22%2C%22is%22%3A%5B%22gte%22%2C500%5D%7D%2C%7B%22field%22%3A%22links_external%22%2C%22is%22%3A%5B%22lte%22%2C300%5D%7D%5D%7D`,
+      });
+    });
+  });
+
+  describe('getOrganicTraffic', () => {
+    it('sends API request with appropriate endpoint query params', async () => {
+      const startDate = '2024-01-29';
+      const endDate = '2024-02-05';
+
+      nock(config.apiBaseUrl)
+        .get('/site-explorer/metrics-history')
+        .query({
+          target: 'test-site.com',
+          date_from: startDate,
+          date_to: endDate,
+          history_grouping: 'weekly',
+          volume_mode: 'average',
+          mode: 'prefix',
+          output: 'json',
+        })
+        .reply(200, organicTrafficMetricsResponse);
+
+      const result = await client.getOrganicTraffic('test-site.com', startDate, endDate);
+      expect(result).to.deep.equal({
+        result: organicTrafficMetricsResponse,
+        fullAuditRef: 'https://example.com/site-explorer/metrics-history?target=test-site.com&date_from=2024-01-29&date_to=2024-02-05&history_grouping=weekly&volume_mode=average&mode=prefix&output=json',
       });
     });
   });