+++ packages/spacecat-shared-http-utils/src/index.js
@@ -12,47 +12,74 @@
 
 import { Response } from '@adobe/fetch';
 
+const HEADER_CONTENT_TYPE = 'content-type';
+const HEADER_ERROR = 'x-error';
+
+const CONTENT_TYPE_JSON = 'application/json';
+
 /**
- * Creates a response with a JSON body. Defaults to 200 status.
- * @param {object} body - JSON body.
- * @param {number} status - Optional status code.
- * @param {object} headers - Optional headers.
+ * Creates a response with a JSON body if the content-type is JSON. Defaults to 200 status.
+ * If a header is already defined and has a different content-type, it is handled accordingly.
+ * @param {object} body - Response body.
+ * @param {number} [status=200] - Optional status code.
+ * @param {object} [headers={}] - Optional headers.
  * @return {Response} Response.
  */
 export function createResponse(body, status = 200, headers = {}) {
-  return new Response(
-    body === '' ? '' : JSON.stringify(body),
-    {
-      headers: { 'content-type': 'application/json; charset=utf-8', ...headers },
-      status,
-    },
-  );
+  let responseBody = body;
+
+  // Check if headers already contain a 'content-type' key
+  if (!headers[HEADER_CONTENT_TYPE]) {
+    // Set content-type to JSON if not already set
+    Object.assign(headers, { [HEADER_CONTENT_TYPE]: `${CONTENT_TYPE_JSON}; charset=utf-8` });
+  }
+
+  // Stringify body if content-type is JSON
+  if (headers[HEADER_CONTENT_TYPE].includes(CONTENT_TYPE_JSON)) {
+    responseBody = body === '' ? '' : JSON.stringify(body);
+  }
+
+  return new Response(responseBody, {
+    headers,
+    status,
+  });
 }
+
 export function ok(body = '') {
   return createResponse(body, 200);
+}
+
+export function created(body) {
+  return createResponse(body, 201);
 }
 
 export function noContent(headers = {}) {
   return createResponse('', 204, headers);
 }
 
+export function found(location, body = '') {
+  return createResponse(body, 302, {
+    Location: location,
+  });
+}
+
 export function badRequest(message = 'bad request', headers = {}) {
   return createResponse({ message }, 400, {
-    'x-error': message,
+    [HEADER_ERROR]: message,
     ...headers,
   });
 }
 
 export function notFound(message = 'not found', headers = {}) {
   return createResponse({ message }, 404, {
-    'x-error': message,
+    [HEADER_ERROR]: message,
     ...headers,
   });
 }
 
 export function internalServerError(message = 'internal server error', headers = {}) {
   return createResponse({ message }, 500, {
-    'x-error': message,
+    [HEADER_ERROR]: message,
     ...headers,
   });
 }