+++ packages/spacecat-shared-http-utils/test/index.test.js
@@ -12,22 +12,50 @@
 /* eslint-env mocha */
 import { expect } from 'chai';
 import {
-  ok, badRequest, notFound, internalServerError, noContent,
+  ok, badRequest, notFound, internalServerError, noContent, found, created, createResponse,
 } from '../src/index.js';
 
+async function testMethod(response, expectedCode, expectedBody) {
+  expect(response.status).to.equal(expectedCode);
+  const responseBody = await (typeof expectedBody === 'string' ? response.text() : response.json());
+  expect(responseBody).to.deep.equal(expectedBody);
+}
+
 describe('HTTP Response Functions', () => {
+  it('createResponse should handle text/plain content type', async () => {
+    const body = 'text body';
+    const headers = { 'content-type': 'text/plain' };
+    const response = await createResponse(body, 200, headers);
+    expect(response.status).to.equal(200);
+    expect(response.headers.get('content-type')).to.equal('text/plain');
+    const responseBody = await response.text();
+    expect(responseBody).to.equal(body);
+  });
+
+  it('createResponse should handle application/json content type', async () => {
+    const body = { success: true };
+    const response = await createResponse(body);
+    expect(response.status).to.equal(200);
+    expect(response.headers.get('content-type')).to.equal('application/json; charset=utf-8');
+    const responseBody = await response.json();
+    expect(responseBody).to.deep.equal(body);
+  });
+
   it('ok should return a 200 OK response with default body', async () => {
     const response = await ok();
-    expect(response.status).to.equal(200);
-    const responseBody = await response.text();
-    expect(responseBody).to.equal('');
+    await testMethod(response, 200, '');
   });
 
   it('ok should return a 200 OK response with custom body', async () => {
-    const response = await ok({ success: true });
-    expect(response.status).to.equal(200);
-    const responseBody = await response.json();
-    expect(responseBody).to.deep.equal({ success: true });
+    const body = { success: true };
+    const response = await ok(body);
+    await testMethod(response, 200, body);
+  });
+
+  it('created should return a 201 CREATED response with custom body', async () => {
+    const body = { success: true };
+    const response = await created(body);
+    await testMethod(response, 201, body);
   });
 
   it('noContent should return a 204 No Content response with default headers', async () => {
@@ -43,6 +71,14 @@ describe('HTTP Response Functions', () => {
     expect(response.status).to.equal(204);
     const responseBody = await response.text();
     expect(responseBody).to.equal('');
+  });
+
+  it('found should return a 302 Found response with correct location and body', async () => {
+    const location = 'https://www.example.com';
+    const response = await found(location);
+
+    expect(response.status).to.equal(302);
+    expect(response.headers.get('Location')).to.equal(location);
   });
 
   it('badRequest should return a 400 Bad Request response with default message and headers', async () => {