+++ packages/spacecat-shared-http-utils/CHANGELOG.md
@@ -1,3 +1,31 @@
+# [@adobe/spacecat-shared-http-utils-v1.3.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-http-utils-v1.3.0...@adobe/spacecat-shared-http-utils-v1.3.1) (2024-05-29)
+
+
+### Bug Fixes
+
+* Refactor createResponse Function to Handle Non-JSON Content-Types ([#244](https://github.com/adobe/spacecat-shared/issues/244)) ([7675f35](https://github.com/adobe/spacecat-shared/commit/7675f35c9ae8644b43b489136da66c21c2d7393b))
+
+# [@adobe/spacecat-shared-http-utils-v1.3.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-http-utils-v1.2.1...@adobe/spacecat-shared-http-utils-v1.3.0) (2024-05-21)
+
+
+### Features
+
+* introduce google client ([#219](https://github.com/adobe/spacecat-shared/issues/219)) ([71eeb64](https://github.com/adobe/spacecat-shared/commit/71eeb64a7efe2037b5d40f0c62668b9eaf41aeb8))
+
+# [@adobe/spacecat-shared-http-utils-v1.2.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-http-utils-v1.2.0...@adobe/spacecat-shared-http-utils-v1.2.1) (2024-05-03)
+
+
+### Bug Fixes
+
+* index.d.ts entry for created ([9bc2988](https://github.com/adobe/spacecat-shared/commit/9bc29889c0d706b7c42a01b504ef5f0be9ba7384))
+
+# [@adobe/spacecat-shared-http-utils-v1.2.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-http-utils-v1.1.4...@adobe/spacecat-shared-http-utils-v1.2.0) (2024-05-03)
+
+
+### Features
+
+* http created method ([#224](https://github.com/adobe/spacecat-shared/issues/224)) ([d540e40](https://github.com/adobe/spacecat-shared/commit/d540e40d3df79e4092698945e0d631b1171a35a7))
+
 # [@adobe/spacecat-shared-http-utils-v1.1.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-http-utils-v1.1.3...@adobe/spacecat-shared-http-utils-v1.1.4) (2024-04-10)