+++ packages/spacecat-shared-rum-api-client/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@adobe/spacecat-shared-rum-api-client-v1.8.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-rum-api-client-v1.8.2...@adobe/spacecat-shared-rum-api-client-v1.8.3) (2024-05-25)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#241](https://github.com/adobe/spacecat-shared/issues/241)) ([85babb2](https://github.com/adobe/spacecat-shared/commit/85babb22b8663f79fe857dead6ecc3bf65674687))
+
+# [@adobe/spacecat-shared-rum-api-client-v1.8.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-rum-api-client-v1.8.1...@adobe/spacecat-shared-rum-api-client-v1.8.2) (2024-05-11)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#228](https://github.com/adobe/spacecat-shared/issues/228)) ([ef2ab41](https://github.com/adobe/spacecat-shared/commit/ef2ab41a9175ec5ba8ec7e2830898e9db01fb2b6))
+
+# [@adobe/spacecat-shared-rum-api-client-v1.8.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-rum-api-client-v1.8.0...@adobe/spacecat-shared-rum-api-client-v1.8.1) (2024-04-24)
+
+
+### Bug Fixes
+
+* use 404 new endpoint ([23ec007](https://github.com/adobe/spacecat-shared/commit/23ec0073540eabf3d460f4705ba4dd40f27f46a0))
+
 # [@adobe/spacecat-shared-rum-api-client-v1.8.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-rum-api-client-v1.7.2...@adobe/spacecat-shared-rum-api-client-v1.8.0) (2024-04-19)